Convert position into floats

+++ src/main/java/org/realityforge/webgl/light/Light.java
@@ -1,7 +1,6 @@
 package org.realityforge.webgl.light;
 
 import javax.annotation.Nonnull;
-import org.joml.Vector3d;
 import org.joml.Vector3f;
 
 final class Light
@@ -9,7 +8,7 @@ final class Light
   @Nonnull
   private final Vector3f _color = new Vector3f( 1.0F, 1.0F, 1.0F );
   @Nonnull
-  private final Vector3d _position = new Vector3d( 8.0, 0.0, -7.0 );
+  private final Vector3f _position = new Vector3f( 8.0F, 0.0F, -7.0F );
 
   @Nonnull
   Vector3f getColor()
@@ -18,7 +17,7 @@ final class Light
   }
 
   @Nonnull
-  Vector3d getPosition()
+  Vector3f getPosition()
   {
     return _position;
   }
